+++ src/app/mod.rs
@@ -63,4 +63,4 @@
 
 pub mod application;
 
-pub use application::Application;
+pub use application::{AppHandler, Application};