+++ src/views/color_dialog.rs
@@ -6,13 +6,13 @@
 //!
 //! Provides a dialog for interactive color selection with live preview.
 
+use super::ViewId;
 use super::button::Button;
 use super::color_selector::ColorSelector;
 use super::dialog::Dialog;
 use super::group::{Group, GroupLike};
 use super::static_text::StaticText;
 use super::window::{Window, WindowLike};
-use super::ViewId;
 use crate::core::command::{CM_CANCEL, CM_OK};
 use crate::core::event::Event;
 use crate::core::geometry::Rect;